refactor: organize trigger panel structure
