+++ internal-packages/workflow-designer-ui/src/editor/properties-panel/trigger-node-properties-panel/ui/common/repository-block.tsx
@@ -1,4 +1,4 @@
-import { GitHubIcon } from "../../../icons";
+import { GitHubIcon } from "../../../../../icons";
 
 export function GitHubRepositoryBlock({
 	owner,
