+++ internal-packages/workflow-designer-ui/src/editor/properties-panel/ui/index.ts
@@ -1,5 +1,4 @@
 export * from "./properties-panel";
 export * from "./node-dropdown";
 export * from "./editable-text";
-export * from "./github-repository-block";
 export * from "./select-repository";
